renames db.ts to dao.ts
